refactor: drop object factory from execution context
