+++ lib/Page.js
@@ -27,6 +27,7 @@ const Tracing = require('./Tracing');
 const {helper, debugError, assert} = require('./helper');
 const {Coverage} = require('./Coverage');
 const {Worker} = require('./Worker');
+const {createJSHandle} = require('./ExecutionContext');
 
 const writeFileAsync = helper.promisify(fs.writeFile);
 
@@ -480,7 +481,7 @@ class Page extends EventEmitter {
    */
   async _onConsoleAPI(event) {
     const context = this._frameManager.executionContextById(event.executionContextId);
-    const values = event.args.map(arg => this._frameManager.createJSHandle(context, arg));
+    const values = event.args.map(arg => createJSHandle(context, arg));
     this._addConsoleMessage(event.type, values);
   }
