+++ lib/Worker.js
@@ -33,7 +33,7 @@ class Worker extends EventEmitter {
     let jsHandleFactory;
     this._client.once('Runtime.executionContextCreated', async event => {
       jsHandleFactory = remoteObject => new JSHandle(executionContext, client, remoteObject);
-      const executionContext = new ExecutionContext(client, event.context, jsHandleFactory, null);
+      const executionContext = new ExecutionContext(client, event.context, null);
       this._executionContextCallback(executionContext);
     });
     // This might fail if the target is closed before we recieve all execution contexts.
